Refactor Phase 2.95 parity tests to consolidate assertions and update module names
