+++ tests/tests_rf/test_topology_snapshots.py
@@ -11,7 +11,7 @@ from ramses_rf import Gateway
 
 from .conftest import TEST_DIR
 
-# Define the log files to use for parity testing
+# Define the log files to use for topology snapshot testing
 LOG_STANDARD = Path(f"{TEST_DIR}/logs/test_phase2_95_topology_parity_packet_log.log")
 LOG_OPENTHERM = Path(
     f"{TEST_DIR}/logs/test_phase2_95_topology_parity_packet_log_OpenTherm.log"
@@ -28,7 +28,7 @@ def suppress_asyncio_warnings(caplog: pytest.LogCaptureFixture) -> None:
 
 @pytest.fixture(params=[LOG_STANDARD, LOG_OPENTHERM], ids=["standard", "opentherm"])
 def log_file_path(request: pytest.FixtureRequest) -> Path:
-    """Provide the packet log file path for topology parity testing."""
+    """Provide the packet log file path for topology snapshot testing."""
     return cast(Path, request.param)
 
 
@@ -55,10 +55,7 @@ async def drain_cqrs_queues(gwy_cqrs: Gateway) -> None:
 @pytest.mark.asyncio
 async def test_topology_builder_snapshot(log_file_path: Path, snapshot: Any) -> None:
     """Test that the async TopologyBuilder yields a consistent schema over time."""
-    # ------------------------------------------------------------------------
-    # STEP 1: Run the Async Pipeline
-    # ------------------------------------------------------------------------
-    # Instantiate the Gateway directly with eavesdrop enabled
+    # Arrange
     from ramses_rf.gateway import GatewayConfig
 
     async_config = GatewayConfig(enable_eavesdrop=True)
@@ -66,6 +63,7 @@ async def test_topology_builder_snapshot(log_file_path: Path, snapshot: Any) -> 
 
     async_gwy = Gateway(None, config=async_config)
 
+    # Act
     await async_gwy.start()
     if async_gwy._engine._transport:
         reader_task = async_gwy._engine._transport.get_extra_info("reader_task")
@@ -73,18 +71,11 @@ async def test_topology_builder_snapshot(log_file_path: Path, snapshot: Any) -> 
             with contextlib.suppress(asyncio.TimeoutError, asyncio.CancelledError):
                 await asyncio.wait_for(reader_task, timeout=30)
 
-    # ------------------------------------------------------------------------
-    # STEP 2: Explicit Queue Drain
-    # ------------------------------------------------------------------------
     await drain_cqrs_queues(async_gwy)
 
     raw_schema = await async_gwy.device_registry.generate_schema()  # type: ignore[attr-defined]
 
     await async_gwy.stop()
 
-    # ------------------------------------------------------------------------
-    # STEP 3: Assert against Snapshot
-    # ------------------------------------------------------------------------
-    # We use syrupy to lock the schema into a snapshot file, so any future
-    # refactors that alter the schema unexpectedly will flag as regressions.
+    # Assert
     assert raw_schema == snapshot
